moved log/reg to separate file. Outside click works
